backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
databases
psycopg2-binary
pydantic
python-dotenv